requirements.txt
ultralytics
numpy
opencv-python
Pillow
moviepy
